First attempt at smushing limitlessled and limitlessled_rf together
