+++ homeassistant/components/limitlessled/manifest.json
@@ -2,7 +2,7 @@
   "domain": "limitlessled",
   "name": "LimitlessLED",
   "documentation": "https://www.home-assistant.io/integrations/limitlessled",
-  "requirements": ["limitlessled==1.1.3"],
+  "requirements": ["limitlessled==1.1.3", "limitlessled_rf==0.11", "lt8900_spi==2.4", "gpiozero==1.5.1"],
   "dependencies": [],
   "codeowners": []
 }
