--- homeassistant/components/limitlessled_rf/__init__.py
@@ -1,1 +1,0 @@
-"""The limitlessled via SPI connected LT8900."""
