--- homeassistant/components/limitlessled_rf/manifest.json
@@ -1,8 +1,0 @@
-{
-  "domain": "limitlessled_rf",
-  "name": "LimitlessLED (Direct RF)",
-  "documentation": "https://www.home-assistant.io/integrations/limitlessled_rf",
-  "requirements": ["limitlessled_rf==0.11", "lt8900_spi==2.4", "gpiozero==1.5.1"],
-  "dependencies": [],
-  "codeowners": []
-}
